+++ otter/check/utils.py
@@ -200,6 +200,21 @@ def logs_event(event_type):
     return event_logger
 
 
+def list_test_files(tests_dir):
+    """
+    Find all of the test files in the specified directory (that is, all ``.py`` files that are not
+    named ``__init__.py``) and return their paths in a sorted list.
+
+    Args:
+        tests_dir (``str``): the path to the tests directory
+
+    Returns:
+        ``list[str]``: the sorted list of all test file paths in ``tests_dir``
+    """
+    return sorted([file for file in glob(os.path.join(tests_dir, "*.py")) \
+            if file != "__init__.py"])
+
+
 def list_available_tests(tests_dir, nb_path):
     """
     Get a list of available questions by searching the tests directory (if present) or the notebook
@@ -215,8 +230,7 @@ def list_available_tests(tests_dir, nb_path):
     get_stem = lambda p: os.path.splitext(os.path.basename(p))[0]
 
     if tests_dir and os.path.isdir(tests_dir):
-        tests = [get_stem(file) for file in glob(os.path.join(tests_dir, "*.py")) \
-            if file != "__init__.py"]
+        tests = map(get_stem, list_test_files(tests_dir))
 
     else:
         if nb_path is None: