+++ otter/test_files/__init__.py
@@ -83,13 +83,14 @@ class GradingResults:
         all_hidden (``bool``): whether all results should be hidden from the student on Gradescope
         tests (``list`` of ``str``): list of test names according to the keys of ``results``
     """
-    def __init__(self, test_files):
+    def __init__(self, test_files, notebook=None):
         self._plugin_data = {}
         self.results = {tf.name: tf for tf in test_files}
         # self.results = {}
         self.output = None
         self.all_hidden = False
         self.pdf_error = None
+        self.notebook = notebook
 
     def __repr__(self):
         return self.summary()